+++ CHANGELOG.md
@@ -5,11 +5,13 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
-## Version 1.5.0 - 2025-01-07
+## Version 1.5.0 - 2025-01-08
 
 ### Fixed
 
 - Redis lookup for websocket via CDS env `cds.requires.redis-websocket`
+- Reworked redis client connection logic
+- Document options
 
 ## Version 1.4.1 - 2024-12-02